Implemented ipconfig command.
It took me a while. ...Actually not. It took me 20 seconds. I swas simply focused on something else  and forgot I have a commit to push.

Can anyone axplain what/who the heck is "the lowliest of cogs"? I've seen this easteregg a few times in this mod, but I have no clue whatsoever, what does it mean.