Tweaked IpLogger to make it more flexible.
I noticed, I said "thrown an exception" on IpCommand.cs, but IpLogger does not really throw an exception, even though someone literally SAID "This is technically an error". Well, time to fix it.

The old PrintServerIps() is now depracted. I'm not sure where is it used, so just please, update to the new one.

After I finish doing, what I've been doing, the only warning left is going to be IpLogger.cs; line 57; use type instead of var. Is it OK to change that?

Also, I fixed a few names-related problems and a couple of typos/grammar mistakes. Is it fine?