Prepared everything to save command system & resolved one TODO.
...And created next one bilion of TODOs in process

This requied me to make PlayerManger static. Not sure, is it good, but I don't care. PlayerManager is the only obvius way to list online players and I could't find ANY other way to locate an instance of PlayerManager. Seriously, the only promising one was to locate a point, where CommandProcessor is invoked, but it leads to nowhere.